travis: postpone 1.9.3 testing for now
There's a problem on 1.9.3 where URI::HTTPS#empty? is being called, and that
isn't actually a method on the URI classes.  Rather than investigate, I'd just
as soon go ahead with the migration to faraday and revisit this later.

For the record, the problem is described in the failing build here:

https://travis-ci.org/orchestrate-io/orchestrate-ruby/jobs/23227310