Migrate purge_key -> purge, new signature
It is unclear what the API is supposed to do when both ?purge=true and If-Match
are specified.  I would *assume* that it behaves as expected, that is, purge if
the If-Match is satisfied, ignore if not, however I don't like to assume.  None
of the other clients' documentation seems to address this, so I have ignored
the combination for now.