migrate list_events to new signature
This is still called 'getEvents' in the Go client, but I figure list_events is
a better name.  It uses a similar function signature, and takes query options
for start/end after/before.  There is a new module, Orchestrate::Helpers, that
contains the logic for that, and I will move the key list function's stuff to
that.