Raise errors on 4xx-class responses
After a lot of thought and a little bit of talking with some of the Orchestrate
people on the community chat, I think that raising an exception on
a non-success response is the Right Thing to do.

I'm separating out the exceptions into subclasses of RequestError (this commit)
and ServiceError (not created yet), under the notion that a RequestError is
under your control, but a ServiceError is not.

This will (erroneously) raise a RequestError on a 5xx-class response, but I'll
fix that next time.