Simplify Client initialization
I went back and forth on this quite a bit.  Ultimately, it comes down to, where
are you stuffing some global variables?  Here, there's just one thing to keep
track of, and it's up to the end-user to decide where to put it.  This behavior
is a bit more like the [ruby redis client](https://github.com/redis/redis-rb).

For the forthcoming Orchestrate::Application class stuff, I plan on providing
a way to do cleaner globals.