SearchResults#offset
I thought about using #drop for this, but that will go ahead and the
enumeration before we can call #take; Then I thought about overloading #drop to
just set the offset and return the SearchResults object, but that breaks
Enumerable.  So, this.  One could argue that perhaps #take should be the same
thing for consistency, but I see #take being the more common use case and want
to optimize for that.